Move mocking transactions endpoint to helpers file
